Move institution serialization to persistenceModel project
