+++ persistence-model/src/main/resources/application-dev.conf
@@ -3,7 +3,7 @@ akka {
   actor {
     warn-about-java-serializer-usage = off
     serializers {
-      institution = "hmda.persistence.institutions.serialization.InstitutionProtobufSerializer"
+      institution = "hmda.persistence.serialization.institutions.InstitutionProtobufSerializer"
     }
     serialization-bindings {
       "hmda.persistence.messages.events.institutions.InstitutionEvents$InstitutionCreated" = institution
