+++ validation/src/main/resources/application.conf
@@ -1,18 +1,19 @@
-akka.persistence.journal.plugin = "akka.persistence.journal.leveldb"
-akka.persistence.query.journal.id = "akka.persistence.query.journal.leveldb"
-# DON'T USE THIS IN PRODUCTION
-akka.persistence.journal.leveldb.native = off
-akka.persistence.journal.leveldb.dir = "target/journal"
-
-akka.persistence.snapshot-store.plugin = "akka.persistence.snapshot-store.local"
-akka.persistence.snapshot-store.local.dir = "target/snapshots"
-
 akka {
-  stdout-loglevel = "OFF"
+  stdout-loglevel = "DEBUG"
   loglevel = "OFF"
   log-dead-letters = off
   log-dead-letters-during-shutdown = off
-  actor.warn-about-java-serializer-usage = off
+  actor {
+    warn-about-java-serializer-usage = off
+
+  }
+  persistence {
+    journal.plugin = "cassandra-journal"
+    snapshot-store.plugin = "cassandra-snapshot-store"
+    query {
+      journal.id = "cassandra-query-journal"
+    }
+  }
 }
 
 hmda {
